BACK-34 added exception handling, added validation
added simple exception handling and validation for employee

+++ service/src/main/java/com/inventory/services/employee/EmployeeServiceImpl.java
@@ -3,6 +3,10 @@ package com.inventory.services.employee;
 import com.inventory.models.Employee;
 import com.inventory.models.Paging;
 import com.inventory.repositories.EmployeeRepository;
+import com.inventory.services.exceptions.EmployeeFieldWrongFormatException;
+import com.inventory.services.exceptions.EmployeeNotFoundException;
+import com.inventory.services.exceptions.EntityNullFieldException;
+import com.inventory.services.validators.EntityValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
@@ -22,14 +26,26 @@ public class EmployeeServiceImpl implements EmployeeService {
     @Autowired
     private PasswordEncoder encoder;
 
+    @Autowired
+    private EntityValidator validator;
+
     @Override
     @Transactional
-    public Employee getEmployee(String id) {
-        return employeeRepository.findById(id).get();
+    public Employee getEmployee(String id) throws EmployeeNotFoundException {
+        if (!validator.validateIdFormatEntity(id, "EM"))
+            throw new EmployeeNotFoundException("id not valid");
+        try {
+            return employeeRepository.findById(id).get();
+        } catch (RuntimeException e) {
+            throw new EmployeeNotFoundException("id not valid");
+        }
     }
 
     @Override
     public Employee login(String email, String password) {
+        boolean isEmailValid = validator.validateEmailFormatEmployee(email);
+        if (!isEmailValid)
+            return null;
         Employee employee;
         try {
             employee = employeeRepository.findByEmail(
@@ -69,11 +85,14 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     @Transactional
-    public List<Employee> getSuperiorList(String superiorId, String name, Paging paging) {
+    public List<Employee> getSuperiorList(String superiorId, String name, Paging paging)
+            throws EmployeeFieldWrongFormatException {
         if (name == null)
             name = "";
         if (superiorId == null)
             superiorId = "null";
+        if (!validator.validateIdFormatEntity(superiorId, "EM"))
+            throw new EmployeeFieldWrongFormatException("Superior Id is not in the right format");
         List<Employee> listOfEmployee;
         if (paging.getSortedType().matches("desc")) {
             listOfEmployee = employeeRepository.findAllBySuperiorIdAndNameContainingIgnoreCase(
@@ -98,25 +117,34 @@ public class EmployeeServiceImpl implements EmployeeService {
 
     @Override
     @Transactional
-    public Employee saveEmployee(Employee employee) {
-        if (employee.getId() == null)
-            employee.setPassword(encoder.encode(employee.getPassword()));
-        if (employee.getSuperiorId().equals("null"))
-            employee.setRole("SUPERIOR");
+    public Employee saveEmployee(Employee employee) throws RuntimeException {
+        employee.setPassword(encoder.encode(employee.getPassword()));
+        employee.setRole(validator.assumeRoleEmployee(employee.getSuperiorId()));
+        String nullFieldEmployee = validator.validateNullFieldEmployee(employee);
+        boolean isSuperiorIdValid = validator.validateIdFormatEntity(employee.getSuperiorId(), "EM");
+        boolean isEmailValid = validator.validateEmailFormatEmployee(employee.getEmail());
+        if (nullFieldEmployee != null)
+            throw new EntityNullFieldException(nullFieldEmployee);
+        else if (!isEmailValid)
+            throw new EmployeeFieldWrongFormatException("Email is not in the right format");
+        else if (!isSuperiorIdValid)
+            throw new EmployeeFieldWrongFormatException("Superior Id is not in the right format");
         else
-            employee.setRole("EMPLOYEE");
-        return employeeRepository.save(employee);
+            return employeeRepository.save(employee);
     }
 
     @Override
     @Transactional
-    public List<String> deleteEmployee(List<String> ids) {
+    public List<String> deleteEmployee(List<String> ids) throws RuntimeException {
         List<String> listOfNotFoundIds = new ArrayList<>();
         for (String id : ids) {
             try {
+                boolean isIdValid = validator.validateIdFormatEntity(id, "EM");
+                if (!isIdValid)
+                    throw new EmployeeFieldWrongFormatException("Id is not in the right format");
                 employeeRepository.deleteById(id);
-            } catch (NullPointerException e) {
-                listOfNotFoundIds.add("id " + id + " not found");
+            } catch (RuntimeException e) {
+                throw new EmployeeNotFoundException("id : " + id + " is not found");
             }
         }
         return listOfNotFoundIds;
